% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPlot.R
\name{clPlot}
\alias{clPlot}
\title{class plot}
\usage{
clPlot(meas = NULL, dpgmm = NULL, nc = 2)
}
\arguments{
\item{meas}{object of class SIRData.}

\item{dpgmm}{object of class clusData.}

\item{nc}{number of column of plot.}
}
\value{
Multiplot panel. Each plot represents a class. The histograms represent the mean
clearness Index distribution of the class and the bars represent the standard deviation
of the class.
}
\description{
Generic function to generate plot of classes.
}
\examples{

data("SIRData_obj")
data("clusData_obj")

# three classes of GHI.
clPlot(SIRData_obj, clusData_obj, 2)

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
