% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusData_obj.R
\docType{data}
\name{clusData_obj}
\alias{clusData_obj}
\title{clusData_obj dataset}
\description{
A dataset for testing use of functions
}
\references{
clusData_obj for runing the example.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
