% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusDirGauss.R
\name{ClusDirGauss}
\alias{ClusDirGauss}
\title{Dirichlet process Gaussian Mixture Model -- DPGMM --}
\usage{
ClusDirGauss(meas, n_iter = NULL, n_burn = NULL)
}
\description{
Generic function of the nonparametric Bayesian Dirichlet-Gaussian clustering.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
