% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPMMclus.R
\name{DPGMMclus}
\alias{DPGMMclus}
\title{DPGMMclus S3 Method}
\usage{
DPGMMclus(obj, n.iter, n.burn, ci, alfa)
}
\arguments{
\item{obj}{object of class SIRData (see \code{\link{SIR_Data}}) or a matrix object containing the clearness index distributions.}

\item{n.iter}{numeric(1) represents the number of iterations}

\item{n.burn}{numeric(1) represents the number of burn-in iterations (ignored iterations)}

\item{ci}{numeric vector that contain the initialization of the indicator variables (the initial assignement of data to clusters).}

\item{alfa}{numeric(1) represents the concentration parameter. Default is 0.5.}
}
\value{
an object of clusData class containing:
\item{cl_seq}{numeric vector represents the class sequence.}
\item{like_par}{list0object represents the parameters of each class. 1st element is the mean (numeric vector), 5th element is the precision matrix
(inverse of the co-variance matrix) of the class.}
\item{lik_con_par}{numeric(1), represents the inferred concentration parameter alf.}
\item{Con_par_sam_seq}{numeric vector, represents the posterior distribution of alf.}
\item{seq_cl_num}{numeric vector, represents the distribution of the class numbers.}
\item{bet_seq}{numeric vector, represents the destribution of the beta parametrer of the precision matrix.}
\item{like_cl}{numeric vector, represents the number of elements of each class.}
\item{calc_time}{numeric(1) represents the computing time consumed.}
}
\description{
Nonparametric Bayesian Dirichlet-Gaussian clustering of daily clearness index distributions. It can be also used to perform any
data clustering of class matrix other than irradiance data of class SIRData.
}
\examples{
data("SIRData_obj")

newClustering <- DPGMMclus(SIRData_obj, n.iter = 1000, n.burn = 500, ci = 5, alfa = 1)
# for class ploting see \\code{\\link{clPlot}}

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
