% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHI_10_min.R
\docType{data}
\name{EHI_10_min}
\alias{EHI_10_min}
\title{EHI_10_min dataset}
\description{
A dataset for testing use of functions
}
\references{
10-min extraterrestrial horizontal solar iradiance time series at Toravere, Estonia
 for the same days as GHI_10_min data.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
