% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extEHIcalc.R
\name{extEHIcalc}
\alias{extEHIcalc}
\title{Extraterrestrial horizontal solar irradiance generation}
\usage{
extEHIcalc(phi = 0, lg = 0, tStep = 60)
}
\arguments{
\item{phi}{numreic(1) between -90 and 90 represents the latitude of the needed location.}

\item{lg}{numreic(1) between -180 and 180 represents the longitude of the needed location.}

\item{tStep}{numreic(1) represents the needed time step in seconds. Example to generate 1-min
data, tStep = 60.}
}
\value{
a matrix object. Each row represents a day.
}
\description{
Generic function to generate the extraterrestrial horizontal solar irradiance data.
It can genarate at any time resolution.
}
\examples{

# EHI generation with 1-min time step in Kenitra, Morocco.
extEHIcalc  <- function(phi = 34, lg = -6, tStep = 60)

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
