% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Solar irradiance data generation}
\usage{
GenData(clusChar = NULL, stp_from = "10 minutes", n_cores = 1, mc = 1000)
}
\arguments{
\item{clusChar}{object of class genData.}

\item{stp_from}{charachter object with values "10 minutes" or "15 minutes". It represents the original time resolution of the solar irradiance data}

\item{n_cores}{numreic(1), is the used number of cores. Defaut is 1. This function uses the parallel calculation for time optimisation.
For WINDOWS users it must be set to 1, since windows platform do not deals with fork() function.}

\item{mc}{numreic(1)represents the number of iterations to choose the most probable paths.
Default is 1000.}
}
\value{
a list object of 1-min solar irradiance data. Each element represents a day.
}
\description{
Generic function to generate high resolution solar irradiance data. It can genarate 1-min data from 10-min and
15-min solar irradiance databases.
}
\examples{
 
# The example and data below are just to give an idea of how the script works and not to judge the performance of the method.

data("genData_obj")
data("GHI_10_min")

# data generation. For WINDOWS users n_cores must be 1.
GHI1min_gen <- GenData(clusChar = genData_obj, stp_from = "10 minutes", n_cores = 1)

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
