% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData_obj.R
\docType{data}
\name{genData_obj}
\alias{genData_obj}
\title{genData_obj dataset}
\description{
A dataset for testing use of functions
}
\references{
genData_obj for runing the example.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
