% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHI_10_min.R
\docType{data}
\name{GHI_10_min}
\alias{GHI_10_min}
\title{GHI_10_min dataset}
\description{
A dataset for testing use of functions
}
\references{
10-min averaged data from GHI_1_min data.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
