% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHI_1_min.R
\docType{data}
\name{GHI_1_min}
\alias{GHI_1_min}
\title{GHI_1_min dataset}
\description{
A dataset for testing use of functions
}
\references{
100 day randomly sampled from 2016 Toravere, Estonia 1-min GHI time-series from BSRN databases.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
