% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parClusGen.R
\name{parClusGen}
\alias{parClusGen}
\title{genData class}
\usage{
parClusGen(measures = NULL, dpgmmRes = NULL, genNewSqClass = FALSE, Xbins = NULL)
}
\arguments{
\item{measures}{object of class SIRData.}

\item{dpgmmRes}{object of class clusData.}

\item{genNewSqClass}{boolean variable. If TRUE the function assigne days to classes based on similar climate approach. Default is 
FALSE.}

\item{Xbins}{numeric(1) used if genNewSqClass is TRUE. It represents the number of bins to create the matrix of the 
daily clearness index distributions.}
}
\value{
an object of class genData
}
\description{
Constructor function of the objects of genData class, needed to perform the generation procedure of hight resolution 
solar irradiance data. It is based on the resulting object dpgmmRes of the clustering procedure of the considered solar 
irradiance data. It also can be used to generate data based on similar climate conditions by setting genNewSqClass as TRUE
and a dpgmmRes object corresponding to the similar climate (other than the considered solar irradiance data).
}
\examples{
 
data("clusData_obj")
data("SIRData_obj")

# creat an object of class genData.
newgenDataObj <- parClusGen(SIRData_obj, clusData_obj)

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
