% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_Data.R
\name{SIR_Dataa}
\alias{SIR_Dataa}
\title{SIRData class}
\usage{
SIR_Data(Ghi_from = NULL, Ehi_from = NULL, Ehi_to = NULL, Xbins = NULL)
}
\arguments{
\item{Ghi_from}{list of the global solar irradiance data with 10-min or 15-min loggin interval. Each component represents a day.
Since the day lengths throughout the year are not equal, the class list is the most suitable.}

\item{Ehi_from}{list of the corresponding extraterrestrial solar irradiance data of Ghi_from with
the same time resolution.}

\item{Ehi_to}{list of the extraterrestrial solar irradiance data of the 1-min resolution.}

\item{Xbins}{numeric(1) represents the number of bins to create the matrix of the daily clearness index distributions.}
}
\value{
an object of class SIRData
}
\description{
Constructor function of the objects of SIRData class needed to perform the clustering procedure of the
daily clearness index distrubtions. Once you create a SIRData object, you no longer need other inputs. All the
rest work will done automatically.
}
\examples{
data("GHI_10_min", "EHI_10_min", "EHI_1_min")

# creat an objet of class SIRData
newSIRDataObj <- SIR_Data(GHI_10_min, EHI_10_min, EHI_1_min, Xbins = 13)

}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
