% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRData_obj.R
\docType{data}
\name{SIRData_obj}
\alias{SIRData_obj}
\title{SIRData_obj dataset}
\description{
A dataset for testing use of functions
}
\references{
SIRData_obj for runing the example.
}
\author{
Azeddine Frimane \email{Azeddine.frimane@uit.ac.ma; Azeddine.frimane@yahoo.com}
}
\keyword{data}
